backend/requirements.txt
Flask==1.0.2
Flask-Cors==3.0.7
Flask-PyMongo==2.2.0
flask-restplus==0.12.1
python-dateutil==2.8.1
pyjwt==1.7.1

